+++ src/commands/builds/log.ts
@@ -1,12 +1,21 @@
 import { ApifyCommand } from '../../lib/command-framework/apify-command.js';
 import { Args } from '../../lib/command-framework/args.js';
-import { error, info } from '../../lib/outputs.js';
+import { info } from '../../lib/outputs.js';
 import { getLoggedClientOrThrow, outputJobLog } from '../../lib/utils.js';
 
 export class BuildsLogCommand extends ApifyCommand<typeof BuildsLogCommand> {
 	static override name = 'log' as const;
 
 	static override description = 'Prints the log of a specific build.';
+
+	static override examples = [
+		{
+			description: 'Print the build log.',
+			command: 'apify builds log <buildId>',
+		},
+	];
+
+	static override docsUrl = 'https://docs.apify.com/cli/docs/reference#apify-builds-log';
 
 	static override args = {
 		buildId: Args.string({
@@ -23,20 +32,11 @@ export class BuildsLogCommand extends ApifyCommand<typeof BuildsLogCommand> {
 		const build = await apifyClient.build(buildId).get();
 
 		if (!build) {
-			error({ message: `Build with ID "${buildId}" was not found on your account.`, stdout: true });
-			return;
+			throw new Error(`Build with ID "${buildId}" was not found on your account.`);
 		}
 
 		info({ message: `Log for build with ID "${buildId}":\n` });
 
-		try {
-			await outputJobLog({ job: build, apifyClient });
-		} catch (err) {
-			// This should never happen...
-			error({
-				message: `Failed to get log for build with ID "${buildId}": ${(err as Error).message}`,
-				stdout: true,
-			});
-		}
+		await outputJobLog({ job: build, apifyClient });
 	}
 }